Add plans window (#249)
* setup window

* feat(assets): Add new background images for pro and local cards

Adds two new background images for the pro and local cards on the
desktop app. The new images provide a more modern and visually
appealing design for these key components of the user interface.

* feat(windows): Increase Plans window size

Increase the default size of the Plans window from 640x360 to
900x900 to provide more space for the content and improve the
user experience.

* feat(settings-button): Add cursor pointer to settings dropdown

Adds a cursor pointer to the settings dropdown menu content to
improve the user experience and indicate that the dropdown is
clickable.

* feat(toolbar): add transparent toolbar for plans view

The changes introduce a new `TransparentToolbar` component that is
displayed when the user is in the "plans" view. This toolbar has a
transparent background and a border-bottom to match the design of the
plans view. The main toolbar logic is updated to render the
`TransparentToolbar` when the user is in the plans view.

* i18n

---------

Co-authored-by: John Jeong <jeeheontransformers@gmail.com>