Enhance Audio Consent UX and Speaker Controls in ListenButton Component (#834)
* feat(listen-button): Improve audio controls and consent handling

This commit introduces several improvements to the `ListenButton` component:

1. Adds a new `CheckIcon` to the audio control buttons to indicate the current state.
2. Handles the case where the user closes the popover without giving consent. In this case, the consent screen is not shown again automatically.
3. Ensures the speaker is muted by default when the recording starts, if the consent screen is shown.
4. Tracks the initial mount state to handle the speaker mute logic correctly.

These changes improve the user experience and make the audio controls more intuitive.

* feat(ongoing-session): add hasShownConsent state and action

The changes in this commit add a new state property `hasShownConsent` and a corresponding action `setHasShownConsent` to the `ongoingSessionStore`. This allows the application to track whether the user has seen and acknowledged the consent dialog for the ongoing session.

* i18n

* feat(listen-button): remove consent settings button

Removes the consent settings button from the listen button component.
This button was not being used and was cluttering the UI, so it has
been removed to simplify the component and improve the user experience.