disable-auto-scroll (#1090)
* feat: Disable auto-scroll functionality

The changes in this commit disable the auto-scroll functionality in the `notes-list.tsx`, `search-list.tsx`, and related components. This was done to improve the user experience by preventing the page from automatically scrolling to the selected item, which can be disruptive and unexpected for the user.

The changes involve commenting out the `useEffect` hooks that were responsible for the auto-scroll behavior, effectively disabling this feature.

* i18n